requirements.txt
aliyun-log-python-sdk
environs
httpx
